==> Notebooks/1. ExtraerTablas.ipynb <==
### Codigo para extraer las tablas de los pdf enviados desde el laboratorio

##### Instala la biblioteca pdfplumber (descomentar en caso de que sea necesario)

## !pip install pdfplumber

##### Importa las bibliotecas necesarias

import pdfplumber
import pandas as pd

##### Establece la ruta del archivo y utiliza pdf plumber para extraer las tablas para luego, mediante pandas unirlas en un solo dataframe que luego guarda en un archivo .xlsx

pdf_path = "RESULTADOS TPM AÑO 2022.pdf"

tables = []
with pdfplumber.open(pdf_path) as pdf:
    for page in pdf.pages:
        extracted_tables = page.extract_tables()
        for table in extracted_tables:
            if table:
                df = pd.DataFrame(table[1:], columns=table[0])
                tables.append(df)

df_final = pd.concat(tables, ignore_index=True)


df_final.to_excel("tablas_2022.xlsx", index=False)

##### Repetir el proceso para cada archivo (uno por año)

pdf_path = "RESULTADOS TPM AÑO 2023.pdf"

tables = []
with pdfplumber.open(pdf_path) as pdf:
    for page in pdf.pages:
        extracted_tables = page.extract_tables()
        for table in extracted_tables:
            if table:
                df = pd.DataFrame(table[1:], columns=table[0])
                tables.append(df)

df_final = pd.concat(tables, ignore_index=True)

df_final.to_excel("tablas_2023.xlsx", index=False)


pdf_path = "RESULTADOS TPM AÑO 2024.pdf"

tables = []
with pdfplumber.open(pdf_path) as pdf:
    for page in pdf.pages:
        extracted_tables = page.extract_tables()
        for table in extracted_tables:
            if table:
                df = pd.DataFrame(table[1:], columns=table[0])
                tables.append(df)

df_final = pd.concat(tables, ignore_index=True)

df_final.to_excel("tablas_2024.xlsx", index=False)
