==> Notebooks/2. LimpiezaDatosLaboratorio.ipynb <==
## Limpieza de datos de analisis de toxina paralizante en moluscos (TPM)

import pandas as pd
import numpy as np

df_2022 = pd.read_excel("tablas_2022.xlsx")

df_2022 = df_2022.apply(lambda x: x.str.replace("\n", " ", regex=True) if x.dtype == "object" else x)

df_2022.head()

df_2022.Resultado.unique()

mapeo = {
    "NO APTO P/CONSUMO" : 2,
    "MENOR A 800": 1,
    "MAYOR A 800" : 2
    
}
df_2022['Resultado_cat'] = df_2022['Resultado'].map(mapeo)
print(df_2022.head(10))

columnas_a_conservar = ["Fecha", "Zona_Clasificada", "Resultado_cat"]

df_relevantes_2022 = df_2022.copy()[columnas_a_conservar]
print(df_relevantes_2022.head())

zonas = df_relevantes_2022.Zona_Clasificada.unique()
for zona in zonas:
    print(F"La zona {np.where(zonas == zona)[0][0]} es: {zona}")

df_relevantes_2022["Zona_Clasificada"] = df_relevantes_2022["Zona_Clasificada"].replace({zonas[0]: zonas[7], zonas[4]: zonas[10],zonas[5]: zonas[11],
                                                                   zonas[8]: zonas[6], zonas[9]: zonas[7]})

zonas = df_relevantes_2022.Zona_Clasificada.unique()
for zona in zonas:
    print(F"La zona {np.where(zonas == zona)[0][0]} es: {zona}")

df_2023 = pd.read_excel("tablas_2023.xlsx")

df_2023 = df_2023.apply(lambda x: x.str.replace("\n", " ", regex=True) if x.dtype == "object" else x)

df_2023.head()

df_2023.Resultado.unique()

mapeo = {
    "MENOR A 800" : 1,
    "NO APTO P/CONSUMO": 2,
    "N.D. (No Detectable)" : 0,
    "NO APTO": 2,
    "NO DETECTABLE": 0,
    "Menor a 800" : 1,
    "N.D.": 0   
}
df_2023['Resultado_cat'] = df_2023['Resultado'].map(mapeo)
print(df_2023.head(10))

df_relevantes_2023 = df_2023.copy()[columnas_a_conservar]
print(df_relevantes_2023.head())

zonas_2023 = df_relevantes_2023.Zona_Clasificada.unique()

for i, zona in enumerate(zonas_2023):
    print(F"La zona {i} es: {zona}")

df_relevantes_2023["Zona_Clasificada"] = df_relevantes_2023["Zona_Clasificada"].replace({
                                                                zonas_2023[0]: zonas[0], zonas_2023[1]: zonas[4],
                                                                zonas_2023[2]: zonas[3], zonas_2023[3]: zonas[2],
                                                                zonas_2023[4]: zonas[5], zonas_2023[5]: zonas[1],
                                                                zonas_2023[6]: zonas[2], zonas_2023[7]: zonas[0],
                                                                zonas_2023[8]: zonas_2023[12], zonas_2023[9]: zonas[4],
                                                                zonas_2023[10]: zonas_2023[12], zonas_2023[11]: zonas_2023[12],
                                                                zonas_2023[13]: zonas_2023[12], zonas_2023[14]: zonas[3],
                                                                zonas_2023[15]: zonas_2023[12], zonas_2023[17]: zonas_2023[12],
                                                                zonas_2023[18]: zonas_2023[12], zonas_2023[22]: zonas_2023[12]})

zonas_2023 = df_relevantes_2023.Zona_Clasificada.unique()

for i, zona in enumerate(zonas_2023):
    print(F"La zona {i} es: {zona}")

conteo_2022 = df_relevantes_2022.Zona_Clasificada.value_counts()
conteo_2023 = df_relevantes_2023.Zona_Clasificada.value_counts()

print("Conteo 2022")
print(conteo_2022)
print("Conteo 2023")
print(conteo_2023)

df_relevantes_2022.info()

zonas_interes = zonas[0:6]

df_filtrado_2022 = df_relevantes_2022[df_relevantes_2022["Zona_Clasificada"].isin(zonas_interes)]

df_filtrado_2022.info()


df_relevantes_2023.info()

valores= df_relevantes_2023.Zona_Clasificada.unique()

for valor in valores:
    print(valor in zonas_interes)

df_relevantes_2023.Zona_Clasificada.value_counts()

df_filtrado_2023 = df_relevantes_2023[df_relevantes_2023["Zona_Clasificada"].isin(zonas_interes)]

df_filtrado_2023.Zona_Clasificada.value_counts()

df_filtrado_2023.info()

df_2024 = pd.read_excel("tablas_2024.xlsx")

# Reemplazar saltos de línea en todo el DataFrame
df_2024 = df_2024.apply(lambda x: x.str.replace("\n", " ", regex=True) if x.dtype == "object" else x)

df_2024.head()

df_2024.Resultado.unique()

mapeo = {
    "N.D. (No Detectable)" : 0,
    "N.D. (NO DETECTABLE)": 0,
    "MENOR A 800" : 1
    
}
df_2024['Resultado_cat'] = df_2024['Resultado'].map(mapeo)
print(df_2024.head(10))

columnas_a_conservar = ["Fecha", "Zona_Clasificada", "Resultado_cat"]

df_relevantes_2024 = df_2024.copy()[columnas_a_conservar]
print(df_relevantes_2024.head())

zonas_2024 = df_relevantes_2024.Zona_Clasificada.unique()

for i, zona in enumerate(zonas_2024):
    print(F"La zona {i} es: {zona}")

df_relevantes_2024["Zona_Clasificada"] = df_relevantes_2024["Zona_Clasificada"].replace({
                                                                zonas_2024[0]: zonas[0], zonas_2024[1]: zonas[4],
                                                                zonas_2024[2]: zonas[3], zonas_2024[3]: zonas[2],
                                                                zonas_2024[4]: zonas[5], zonas_2024[5]: zonas[1],
                                                                zonas_2024[6]: zonas[4], zonas_2024[7]: zonas[3],
                                                                zonas_2024[8]: zonas[2], zonas_2024[9]: zonas[5],
                                                                zonas_2024[10]: zonas[0], zonas_2024[11]: zonas[0],
                                                                zonas_2024[12]: zonas[1], zonas_2024[13]: zonas[1]})

zonas_2024 = df_relevantes_2024.Zona_Clasificada.unique()

for i, zona in enumerate(zonas_2024):
    print(F"La zona {i} es: {zona}")

df_filtrado_2024 = df_relevantes_2024[df_relevantes_2024["Zona_Clasificada"].isin(zonas_interes)]

df_filtrado_2024.Zona_Clasificada.value_counts()

df_filtrado_2024.info()

df_unido = pd.concat([df_filtrado_2022, df_filtrado_2023, df_filtrado_2024])

df_unido = df_unido.reset_index(drop=True)

print(df_unido.info())

df_unido.to_excel("limpio_unido.xlsx", index=False)
