==> Notebooks/3. Integracion_datos_TPM.ipynb <==
## Lectura e integración de datos de análisis de toxina paralizante en moluscos (TPM)

**Importación de librerias**

import pandas as pd
import numpy as np

**Lectura del primer dataset**

df_antiguo = pd.read_excel("antiguo_unido.xlsx")
df_antiguo.head()

**Lectura de nuevos datos**

df_nuevo = pd.read_excel("limpio_unido.xlsx")
df_nuevo.head()

**Unión de ambos sets de datos**

df_concatenado = pd.concat([df_antiguo, df_nuevo], ignore_index=True)
df_concatenado.info()

df_concatenado.head()

**Ordenamiento del set de datos y re-indexación para numerar las filas nuevamente.**

df_ordenado = df_concatenado.sort_values('Fecha').reset_index(drop=True)
df_ordenado.head()

**Agrupacion por fecha tomando el valor maximo de categoria (0: no detectable, 1: apto p/consumo, 2: no apto para consumo)**

df_resumen = df_ordenado.groupby('Fecha', as_index=False)['Resultado_cat'].max()
df_resumen.head()

**Se completan las fechas que no tienen analisis para tener un rango temporal diario**

# Crear un rango de fechas diario
rango_fechas = pd.date_range(start='2021-01-01', end=df_resumen['Fecha'].max(), freq='D')

# Reindexar usando ese rango
df_completo = df_resumen.set_index('Fecha').reindex(rango_fechas).reset_index()

# Renombrar la columna de fecha
df_completo = df_completo.rename(columns={'index': 'Fecha'})

df_completo.head(20)

**Visualización de registros por cada categoria**

df_completo["Resultado_cat"].value_counts()

Se rellenan los vacios (donde no hay analisis) tomando el valor del analisis inmediato siguiente

# Rellenar 'Resultado_cat' usando forward-fill (ffill)
df_completo["Resultado_cat"] = df_completo["Resultado_cat"].ffill()

# Rellenar los nulos restantes (al inicio) con 0
df_completo["Resultado_cat"] = df_completo["Resultado_cat"].fillna(0)

print("Columna 'Resultado_cat' rellenada.")
df_completo.head()

**Visualizacion registros por categoria luego de rellenar los nulos**

df_completo["Resultado_cat"].value_counts()

df_completo['Resultado_cat'] = df_completo['Resultado_cat'].astype('int')
df_completo['Resultado_cat'] = df_completo['Resultado_cat'].astype('category')

df_completo.head()

**Exportación del set de datos final**

df_completo.to_excel("completo_TPM.xlsx", index=False)
