==> Notebooks/4. Integracion_datos_meteorologicos.ipynb <==
## Lectura e intregración de datos meteorológicos

**Importación de librerias**

import pandas as pd

**Lectura de ambos sets de datos (misma rango temporal distintas variables)**

df_1 = pd.read_csv("POWER LARC 1.csv")
df_1.head()

df_2 = pd.read_csv("POWER LARC 2.csv")
df_2.head()

**Unión de ambos sets de datos**

df_unido = pd.merge(df_2, df_1, on=['YEAR', 'MO', 'DY'], how='inner')
df_unido.head()


**Tratamiento de las columnas de fechas para que sean compatibles a los datos de toxinas**

df_unido['Fecha'] = pd.to_datetime(
    df_unido[['YEAR', 'MO', 'DY']].rename(columns={'YEAR': 'year', 'MO': 'month', 'DY': 'day'})
)
df_unido.head()

df_unido= df_unido.drop(columns=['YEAR', 'MO', 'DY'])

# Reordenar columnas, poniendo 'Fecha' al principio
columnas = ['Fecha'] + [col for col in df_unido.columns if col != 'Fecha']
df_unido = df_unido[columnas]
df_unido.head()

print("Calculando features de medias móviles...")

df_unido = df_unido.sort_values(by='Fecha')

ventanas = [7, 14, 30]

vars_mean = [
    'T2M', 'T2M_MAX', 'T2M_MIN',
    'ALLSKY_SFC_PAR_TOT', 'ALLSKY_KT',
    'WS2M'
]

vars_sum = [
    'PRECTOTCORR'
]

for v in ventanas:
    for var in vars_mean:
        col_name = f'{var}_MA_{v}d'
        df_unido[col_name] = df_unido[var].shift(1).rolling(window=v).mean()
        
    for var in vars_sum:
        col_name = f'{var}_SUM_{v}d'
        df_unido[col_name] = df_unido[var].shift(1).rolling(window=v).sum()

print("Rellenando NaNs iniciales...")
df_unido = df_unido.bfill()

print("¡Features de medias móviles creadas!")
df_unido.head()

df_unido = df_unido[df_unido['Fecha'] >= '2021-01-01']
df_unido.head()

**Exportación**

df_unido.to_csv('POWER_LARC_unido.csv', index=False)
